Make html and js file
This commit establishes the html and javascript files for the project.
The javascript folder contains some psuedocode that has all of the very,
VERY basic functions and steps that will need to be made for the
project. I probably missed a couple of steps here and there, but this is
a good outline for what needs to be done.