Add script to HTML, add choice functions
This commit adds the actual game script to the HTML file. While the
element was in the file, the actual script that contains all of the
necessary information to run the game was not in yet.

This commit also has additions to the javascript file, notably, the
functions to get a random CPU selection, as well as a prompt to have
the player select their choice. I have a feeling these are going to need
some massive changes before the script is finished, but it works for
now.

That's enough javascript for the night.