Add win and lose functions, play round functions
This commit adds functions that define how a user wins or loses against
the computer, as well automating how a round will play out.

Additionally, it adds a counter to keep the score between the computer
and the user. There are some things regarding returning values and
having values that will be used within the parenthesis that I'm still a
little lost on, so I will definitely need to review those concepts
again.

For now, it works...