Add game function
This commit adds the game function, which runs five rounds. After five
rounds, the browswer displays an alert and lets the user know if they
won or lost with the final score at the end.

There's plenty I still don't understand completely, but I know I will be
returning to this project.